\name{summary.gdina}

\alias{summary.gdina}

\title{
  Summary Method for Objects of Class gdina
}

\description{
  S3 method to summarize objects of the class \code{gdina}.  
}

\usage{
  \method{summary}{gdina}(object, rdigits = 4 ,  \dots)
}

\arguments{
  \item{object}{a required object of class \code{gdina}, obtained
    from a call to the function \code{\link{gdina}}.  }
  
  \item{rdigits}{number of digits after decimal separator to display.  }    
            
  \item{\dots}{Optional parameters to be passed to or from other
    methods will be ignored.  }
}

%\details{
%  xxxx
%}

%\value{
%  xxxx
%}

%\author{
%  Alexander Robitzsch <a.robitzsch@bifie.at>, ...
%}


\seealso{
    This function uses object generated from the \code{\link{gdina}} function.
        }
        
%\examples{
%## see examples for gdina function
%}

\keyword{methods}
\keyword{summary}
