\name{summary.gdm}

\alias{summary.gdm}

\title{
  Summary Method for Objects of Class \code{gdm}
}

\description{
  S3 method to summarize objects of the class \code{gdm}.  
}

\usage{
  \method{summary}{gdm}(object, \dots)
}

\arguments{
  \item{object}{a required object of class \code{gdm}}
  \item{\dots}{Optional parameters to be passed to or from other
    methods will be ignored.  }  
}

%\details{
%  xxxx
%}

%\value{
%  xxxx
%}

%\author{
%  Alexander Robitzsch
%}


\seealso{
See \code{\link{gdm}} for examples.
        }
        
%\examples{
%}

\keyword{methods}
\keyword{summary}
