\name{logLik}
\alias{logLik.din}
\alias{logLik.gdina}
\alias{logLik.gdm}

\title{Extract Log-Likelihood}
\description{
  Extracts the log-likelihood from either \code{din}, \code{gdina}
  or \code{gdm} objects.
}
\usage{
\method{logLik}{din}(object, \dots)

\method{logLik}{gdina}(object, \dots)

\method{logLik}{gdm}(object, \dots)
}

\arguments{
  \item{object}{ An object inheriting from either class \code{din}, 
  class \code{gdina}, or class \code{gdm}. 
  	}
  \item{\dots}{Additional arguments}
}

\keyword{methods}
\keyword{logLik}

